pyproject.toml
[build-system]
requires = ["setuptools>=77.0.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nirvana-ai"
version = "1.0.2"
description = "An LLM-powered semantic data analytics programming framework."
readme = "README.md"
authors = [{ name = "Junhao Zhu", email = "zhujunhao@zju.edu.cn" }]
license = { file = "LICENSE" }
requires-python = ">=3.10, <3.14"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3"
]
dependencies = [
    "anyio",
    "bm25s>=0.2.13",
    "chromadb>=1.0.15",
    "numpy>=2.2.4",
    "openai>=1.75.0",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pyarrow>=19.0.1",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.9",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["nirvana", "nirvana.*"]
exclude = ["tests", "tests.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::ResourceWarning",
    "ignore::UserWarning",
]

[project.urls]
repository = "https://github.com/JunHao-Zhu/nirvana"

[tool.ruff]
include = ["nirvana/**/*.py", "tests/**/*.py"]


line-length = 120
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "RUF", # ruff-specific rules
    "Q",   # flake8-quotes
]

ignore = [
    "B027",  # Allow non-abstract empty methods in abstract base classes
    "FBT003",# Allow boolean positional values in function calls
    "C901",  # Ignore complexity checking
    "E501",  # Ignore line length errors (handled by formatter)
    "UP035", # Allow python typing modules
    "RUF005", # Allow using + operator to concatenate collections
    "B904", # Allow raise custom exceptions in except blocks
    "F403", # Allow wildcard imports
    "E721", # Allow using == to compare with type
    "UP031", # Allow percent format
    "RUF022", # Allow unsorted __all__ value
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
